requirements.txt
pip-requirements
langchain
langchain-community
langchain-mcp-adapters
langchain-google-genai
langgraph
python-dotenv
requests

